classstudent 하던거랑 review

+++ src/main/java/mat/mat_t/form/InstructorReviewForm.java
@@ -14,8 +14,8 @@ public class InstructorReviewForm {
     private String reviewContent;
 
     @NotNull
-    private int loginId;
+    private Long loginId;
 
     @NotNull
-    private int classId;
+    private Long classId;
 }
